backend/requirements.txt
Flask==3.1.2
flask-cors==5.0.0
python-dotenv==1.1.1
google-generativeai==0.8.3
requests==2.32.5
beautifulsoup4==4.12.3
lxml==4.9.4
urllib3==2.5.0
Werkzeug==3.1.3
click==8.1.8
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2